Fix bug in flag masking and add enum flag to enum type defs

+++ compiler/ballerina-lang/src/main/java/org/wso2/ballerinalang/util/Flags.java
@@ -71,8 +71,8 @@ public class Flags {
 
     public static final long ENUM = CONFIGURABLE << 1;                          // 32
 
-    public static int asMask(Set<Flag> flagSet) {
-        int mask = 0;
+    public static long asMask(Set<Flag> flagSet) {
+        long mask = 0;
         for (Flag flag : flagSet) {
             switch (flag) {
                 case PUBLIC:
